refactor(notebook): extract control and shell execution owner
